+++ src/builds/VisionPlayer.dev.js
@@ -8,7 +8,7 @@
  * @requires src/visualizer/frequency/VisualizerFrequency
  * @requires src/util/Debug
  * @author   Frank Kudermann - alphanull
- * @version  1.2.2
+ * @version  1.2.3
  * @license  MIT
  */